+++ artemis-fluid/artemis-fluid-core-utility-sources/src/test/java/com/artemis/E.java
@@ -2,7 +2,7 @@ package com.artemis;
 
 /**
  * Placeholder for utility classes that depend on E.
- *
+ * <p>
  * Serves as a placeholder for testing fluid utility classes. This file is never deployed with the utility classes,
  * instead {@link com.artemis.FluidGenerator} generates a user specific version. See artemis-fluid-core pom for
  * build details.
@@ -10,4 +10,11 @@ package com.artemis;
  * @author Daan van Yperen
  */
 public class E {
+    public static E E(int entityId) {
+        throw new RuntimeException("You shouldn't be using this class.");
+    }
+
+    public static E E(Entity entity) {
+        throw new RuntimeException("You shouldn't be using this class.");
+    }
 }