+++ artemis-fluid/artemis-fluid-core-utility-sources/src/test/java/com/artemis/FluidIteratingSystem.java
@@ -1,0 +1,54 @@
+package com.artemis;
+
+import com.artemis.managers.TagManager;
+import com.artemis.systems.EntityProcessingSystem;
+import com.artemis.systems.IteratingSystem;
+
+/**
+ * Iterates over {@link EntitySubscription} member entities
+ * using fluid entities.
+ * <p>
+ * Use this when you need to process entities matching an {@link Aspect}.
+ *
+ * Provides some convenience methods to resolve entities.
+ *
+ * @author Daan van Yperen
+ */
+public abstract class FluidIteratingSystem extends IteratingSystem {
+
+    public FluidIteratingSystem(Aspect.Builder aspect) {
+        super(aspect);
+    }
+
+    @Override
+    protected void process(int id) {
+        process(E.E(id));
+    }
+
+    protected abstract void process(E e);
+
+    /**
+     * Return all entities matching aspect.
+     * Calling the aspect builder is relatively expensive but should be fine outside tight loops.
+     */
+    protected EBag allEntitiesMatching(Aspect.Builder scope) {
+        return new EBag(world.getAspectSubscriptionManager().get(scope).getEntities());
+    }
+
+    /**
+     * Return all entities matching a class.
+     * Calling the aspect builder is relatively expensive but should be fine outside tight loops.
+     */
+    protected EBag allEntitiesWith(Class<? extends Component> scope) {
+        return new EBag(world.getAspectSubscriptionManager().get(Aspect.all(scope)).getEntities());
+    }
+
+    /**
+     * @param tag tag to match. Requires TagManager.
+     * @return Entity matching tag, or {@code null} if missing.
+     */
+    protected E entityWithTag(String tag) {
+        final Entity entity = world.getSystem(TagManager.class).getEntity(tag);
+        return entity != null ? E.E(entity) : null;
+    }
+}